Migrate to using golang buildsystem
Upstream direnv is starting to organised as go package. Migrate
debian/rules to use the golang buildsystem that sets the necessary
environment variables to build a package using go.

The golang buildsystem doesn't quite deal with building and installing
the manpages correctly. Override some of the dh helper rules to do the
right thing for the manpages.