Revert the use of golang package buildsystem
Upstream is including dependencies in vendor folder which isn't
supported by the golang debhelpers without individually packaging the
dependencies.

Requiring to package each dependency is an overkill for direnv so
reverting to the use of regular debhelpers and GOPATH.

This change also enables package tests that had to be disabled in the
previous package release.