debian/control: Annotate test dependencies with nocheck profile
As part of the build, direnv runs tests on different shell to ensure
there are no regressions. These tests bring additional dependencies
which are not required for the build itself.

Annotate the test dependencies with "nocheck" build profile to
reduce dependency installation during build only. This helps speed up
builds during package development and is a worthy goal in itself.